pyproject.toml
[project]
name = "openai-agent-sdk-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.33.1",
    "openai-agents[litellm]>=0.0.17",
    "python-dotenv>=1.1.0",
]
